requirements.txt
# 猫厕所监控系统 - 依赖清单

# 核心依赖
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 目标检测
ultralytics>=8.0.0

# 目标追踪
filterpy>=1.4.5
scikit-learn>=1.3.0
scikit-image>=0.21.0

# 数据处理
pandas>=2.0.0

# 数据库
sqlalchemy>=2.0.0

# Web框架
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# 配置管理
pyyaml>=6.0

# 日志
colorlog>=6.7.0

# 工具
tqdm>=4.65.0
python-dateutil>=2.8.0
requests>=2.31.0

# 可选依赖（用于RTSP流）
ffmpeg-python>=0.2.0

# 开发依赖
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
